requirements.txt
asgiref==3.4.1
certifi==2020.12.5
charset-normalizer==2.0.8
Django==3.2.6
django-cors-headers==3.10.0
djangorestframework==3.12.4
idna==3.3
PyJWT==2.3.0
python-decouple==3.4
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.7
wincertstore==0.2
